fix: (HDS-2686) more shared utilities use in dropdownComponents
