+++ packages/react/src/components/dropdownComponents/modularOptionList/utils.ts
@@ -1,4 +1,4 @@
-import React, { ReactElement, ReactNode } from 'react';
+import { ReactElement, ReactNode } from 'react';
 
 import {
   ModularOptionListData,
@@ -10,20 +10,15 @@ import {
   GroupInProps,
 } from './types';
 import { getChildrenAsArray } from '../../../utils/getChildren';
-import { ChangeEvent } from '../../dataProvider/DataContext';
-import { eventTypes } from './events';
-import { createElementIds, ElementIdsConfig } from '../shared/utils/elementIds';
+import { createElementIds, ElementIdsConfig, createOnClickListener, createInputOnChangeListener } from '../shared';
 
 // Configuration for ModularOptionList element IDs
 const modularOptionListElementIds: ElementIdsConfig = {
   list: true,
 };
 
-type DomHandlerProps = {
-  id: string;
-  type?: string;
-  trigger: (event: ChangeEvent) => void;
-};
+// Re-export DOM handler creators for backward compatibility
+export { createOnClickListener, createInputOnChangeListener };
 
 export type OptionIterator = (
   option: Option,
@@ -33,28 +28,6 @@ export type OptionIterator = (
 ) => Option | undefined;
 
 export const DROPDOWN_MENU_ITEM_HEIGHT = 52;
-
-export function createOnClickListener(props: DomHandlerProps) {
-  const { id, type = eventTypes.click, trigger } = props;
-  return {
-    onClick: (originalEvent: React.MouseEvent) => {
-      trigger({ id, type, payload: { originalEvent } });
-    },
-  };
-}
-
-export function createInputOnChangeListener(props: DomHandlerProps) {
-  const { id, type = eventTypes.change, trigger } = props;
-  return {
-    onChange: (originalEvent: React.ChangeEvent<HTMLInputElement>) => {
-      trigger({
-        id,
-        type,
-        payload: { value: originalEvent.currentTarget.value, originalEvent },
-      });
-    },
-  };
-}
 
 export function getOptionIndex(options: OptionInProps[], option: OptionInProps): number {
   return options.findIndex(
