+++ packages/react/src/components/dropdownComponents/select/utils.ts
@@ -1,9 +1,5 @@
-import React from 'react';
-
 import { SelectData, SelectMetaData, FilterFunction } from './types';
 import { Group, Option, ModularOptionListData, ModularOptionListProps } from '../modularOptionList/types';
-import { ChangeEvent } from '../../dataProvider/DataContext';
-import { eventTypes } from './events';
 import {
   iterateAndCopyGroup,
   getGroupLabelOption,
@@ -11,8 +7,12 @@ import {
   propsToGroups,
   getSelectedOptions,
 } from '../modularOptionList/utils';
-import { ElementIdsConfig } from '../shared/utils/elementIds';
-import { createGenericElementIds } from '../shared';
+import {
+  ElementIdsConfig,
+  createGenericElementIds,
+  createOnClickListener,
+  createInputOnChangeListener,
+} from '../shared';
 
 // Configuration for Select component element IDs
 const selectElementIds: ElementIdsConfig = {
@@ -42,11 +42,8 @@ export {
   defaultFilter,
 } from '../modularOptionList/utils';
 
-type DomHandlerProps = {
-  id: string;
-  type?: string;
-  trigger: (event: ChangeEvent) => void;
-};
+// Re-export DOM handler creators for backward compatibility
+export { createOnClickListener, createInputOnChangeListener };
 
 export type OptionIterator = (
   option: Option,
@@ -56,28 +53,6 @@ export type OptionIterator = (
 ) => Option | undefined;
 
 export const DROPDOWN_MENU_ITEM_HEIGHT = 52;
-
-export function createOnClickListener(props: DomHandlerProps) {
-  const { id, type = eventTypes.click, trigger } = props;
-  return {
-    onClick: (originalEvent: React.MouseEvent) => {
-      trigger({ id, type, payload: { originalEvent } });
-    },
-  };
-}
-
-export function createInputOnChangeListener(props: DomHandlerProps) {
-  const { id, type = eventTypes.change, trigger } = props;
-  return {
-    onChange: (originalEvent: React.ChangeEvent<HTMLInputElement>) => {
-      trigger({
-        id,
-        type,
-        payload: { value: originalEvent.currentTarget.value, originalEvent },
-      });
-    },
-  };
-}
 
 export function countVisibleOptions(groups: SelectData['groups']): number {
   let count = 0;
