+++ packages/react/src/components/dropdownComponents/shared/index.ts
@@ -1,4 +1,6 @@
 export * from './utils/screenReader';
 export * from './utils/elementIds';
+export * from './utils/domHandlers';
+export * from './utils/dataUpdaterHelpers';
 export * from './events';
 export * from './hooks';
